Fix horrible line wrapping in <ul>
Not totally sure what this style was doing copied over from the main
site. But empirically it totally breaks a <ul> with <p> inside the <li>,
and while I'm not sure that's actually valid HTML, docbook generates it
a ton, so we need to not totally explode on every single unordered list.