Establish infra for examples to dynamically added to docs
In order to make sure examples are current and run correctly, we
are going to make the examples part of a "doc test suite" where
the Hack type checker and HHVM will run them and ensure the output
is consistent.

This is the beginning of the infraastructure for that where we
include one exmaple dynamically.

Reference #482